Added unrealised gains, handled edge cases.

+++ backend/models/models.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import List
 from django.db import models
 from django.db.models import Model
@@ -77,7 +78,7 @@ class TradeModel(Model):  # type: ignore
     article_evidence = models.ManyToManyField(ArticleModel, blank=True)
 
     @staticmethod
-    def create_typed(user: UserModel, stock: StockModel, units_change: float, balance_change: float, time: datetime, text_evidence: str, article_evidence: List[ArticleModel]) -> 'TradeModel':
+    def create_typed(user: UserModel, stock: StockModel, units_change: Decimal, balance_change: Decimal, time: datetime, text_evidence: str, article_evidence: List[ArticleModel]) -> 'TradeModel':
         trade = TradeModel.objects.create(
             user=user, stock=stock, units_change=units_change, balance_change=balance_change, time=time, text_evidence=text_evidence)
         trade.article_evidence.set(article_evidence)
